% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_unify_latin.R
\name{get_table_char_unicode}
\alias{get_table_char_unicode}
\title{Replace latin letters with extensions with their respective base forms}
\usage{
get_table_char_unicode()
}
\value{
\code{get_table_char_unicode} - returns a ...
}
\description{
\code{get_table_char_unicode} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
