% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_blur_url.R
\docType{data}
\name{pattern_reg_url}
\alias{pattern_reg_url}
\title{Dictionary for replacing latin letters with extensions with their respective base forms}
\format{
Named character vector where the names are regular expressions and values are regular latin characters without extensions
}
\usage{
data(pattern_reg_url)
}
\description{
Data from ...
}
\keyword{datasets}
