% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%s+\%}
\alias{str_c}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{stringi}{\code{\link[stringi:operator_add]{\%s+\%}}}

  \item{stringr}{\code{\link[stringr]{str_c}}}
}}

