% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_blur_url.R
\name{str_blur_url}
\alias{str_blur_url}
\title{Filter Rows That Contain Markers}
\usage{
str_blur_url(
  .str,
  .pattern_reg_url = pattern_reg_url,
  .replacement_url = "<URL>"
)
}
\arguments{
\item{.str}{...}

\item{.pattern_reg_url}{...}

\item{.replacement_url}{...}
}
\value{
\code{str_blur_url} - returns a ...
}
\description{
\code{str_blur_url} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_blur_url EXAMPLE:

str_blur_url(c(
  "The 2020 United States presidential election was the 59th quadrennial ",
  "presidential election, held on Tuesday, November 3, 2020"
))
}
