% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_convert_html.R
\name{str_convert_html}
\alias{str_convert_html}
\title{Filter Rows That Contain Markers}
\usage{
str_convert_html(.str)
}
\arguments{
\item{.str}{...}
}
\value{
\code{str_convert_html} - returns a ...
}
\description{
\code{str_convert_html} - Convert all Emojis to some ...
}
\examples{
## str_convert_html EXAMPLE:

str_convert_html("&gt; &amp; &lt;")
}
