% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_ascii.R
\name{str_convert_nonascii}
\alias{str_convert_nonascii}
\title{Unicode to ASCII transliteration}
\usage{
str_convert_nonascii(.str, .parse_html_chars = TRUE, .repl_no_trans = "")
}
\arguments{
\item{.str}{Character vector to be transliterated}

\item{.repl_non_ascii}{Value to replace non-transliterable characters with. Set to \code{NULL} to keep them.}
}
\value{
\code{str_convert_nonascii} - Returns the transliterated character vector
}
\description{
\code{str_convert_nonascii} - Converts Unicode characters to their best ASCII representation using ICU Transliterators and some basic transformations.
}
\examples{
## str_convert_nonascii EXAMPLE:

str_convert_nonascii(.str="¿Ŧêśť &amp; 法?")
}
