% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_describe_emoji.R
\name{str_describe_emoji}
\alias{str_describe_emoji}
\title{Filter Rows That Contain Markers}
\usage{
str_describe_emoji(.str, .resolution = "name")
}
\arguments{
\item{.str}{...}

\item{.resolution}{... \code{c("name", "subgroup", "group")}}
}
\value{
\code{str_describe_emoji} - returns a ...
}
\description{
\code{str_describe_emoji} - Convert all Emojis to some ...
}
\examples{
## str_describe_emoji EXAMPLE:

example_emoji <- "😀 😆 😡 💀"
str_describe_emoji(example_emoji, .resolution="name")
str_describe_emoji(example_emoji, .resolution="subgroup")
str_describe_emoji(example_emoji, .resolution="group")
}
