% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_describe_numbers.R
\name{str_describe_numbers}
\alias{str_describe_numbers}
\title{\code{str_describe_numbers} - Replace all cardinal or ordinal numbers in a character vector with generic tags.}
\usage{
str_describe_numbers(.str, .lang = "en", .ordinal = TRUE)
}
\arguments{
\item{.str}{...}

\item{.lang}{...}

\item{.ordinal}{...}
}
\value{
\code{str_describe_numbers} - returns a ...
}
\description{
\code{str_describe_numbers} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_describe_numbers EXAMPLE:

str_describe_numbers(c(
  "The 2020 United States presidential election",
  "was the 59th quadrennial presidential election",
  "held on Tuesday, November 3, 2020"
))
}
