% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_remove_reply_mentions.R
\name{str_remove_reply_mentions}
\alias{str_remove_reply_mentions}
\title{Remove all \code{@mentions} from the beginning of strings.}
\usage{
str_remove_reply_mentions(.str)
}
\arguments{
\item{.str}{...}
}
\value{
\code{str_remove_reply_mentions} - returns a ...
}
\description{
\code{str_remove_reply_mentions} - Remove all \code{@mentions} from the beginning of strings. This can be useful for cleaning tweets because the texts of replies start with the \code{screen_names} of all users further down the reply tree.
}
\examples{
## str_remove_reply_mentions EXAMPLE:

str_remove_reply_mentions(c(
  "@test test!",
  "@test @test test test!"
))
}
