% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_ascii.R
\name{str_to_ascii}
\alias{str_to_ascii}
\title{Replace latin letters with extensions with their respective base forms}
\usage{
str_to_ascii(.str, .repl_non_ascii = "")
}
\arguments{
\item{.str}{...}

\item{.repl_non_ascii}{..}
}
\value{
\code{str_to_ascii} - returns a ...
}
\description{
\code{str_to_ascii} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_to_ascii EXAMPLE:

str_to_ascii(.str="Ŧêśť")
}
