% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_unify_latin.R
\name{str_unify_char_latin}
\alias{str_unify_char_latin}
\title{Replace latin letters with extensions with their respective base forms}
\usage{
str_unify_char_latin(
  .str,
  .table_char = klartext::table_char_latin,
  .col_pattern = pattern,
  .col_replacement = replacement
)
}
\arguments{
\item{.str}{...}

\item{.table_char}{..}

\item{.col_pattern}{..}

\item{.col_replacement}{..}
}
\value{
\code{str_unify_char_latin} - returns a ...
}
\description{
\code{str_unify_char_latin} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_unify_char_latin EXAMPLE:

str_unify_char_latin(.str="Ŧêśť")
}
