% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_normalize_punctuation.R
\name{str_unify_punctuation}
\alias{str_unify_punctuation}
\title{Filter Rows That Contain Markers}
\usage{
str_unify_punctuation(.str)
}
\arguments{
\item{.str}{...}
}
\value{
\code{str_unify_punctuation} - returns a ...
}
\description{
\code{str_unify_punctuation} - Convert all Emojis to some ...
}
\examples{
## str_unify_punctuation EXAMPLE:

str_unify_punctuation(
  "fdsfsd sd  f\n fdsfsd, <TAG> test!"
)
}
