% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_ascii.R
\docType{data}
\name{table_char_latin}
\alias{table_char_latin}
\title{Dictionary for replacing latin letters with extensions with their respective base forms}
\format{
Named character vector where the names are regular expressions and values are regular latin characters without extensions
}
\usage{
data(table_char_latin)
}
\description{
Data from ...
}
\keyword{datasets}
