% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_blur_numbers.R
\name{str_blur_numbers}
\alias{str_blur_numbers}
\title{Filter Rows That Contain Markers}
\usage{
str_blur_numbers(
  .str,
  .rep_cardinal = "<NUM_CARDI>",
  .rep_ordinal = "<NUM_ORDI>",
  .lang = "en",
  .excl_substr = TRUE,
  .excl_hash_at = TRUE,
  .relax_space = TRUE
)
}
\arguments{
\item{.str}{...}

\item{.rep_cardinal}{...}

\item{.rep_ordinal}{...}

\item{.lang}{...}

\item{.excl_substr}{...}

\item{.excl_hash_at}{...}

\item{.relax_space}{...}
}
\value{
\code{str_blur_numbers} - returns a ...
}
\description{
\code{str_blur_numbers} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_blur_numbers EXAMPLE:

str_blur_numbers(c(
  "The 2020 United States presidential election",
  "was the 59th quadrennial presidential election",
  "held on Tuesday, November 3, 2020"
))
}
