% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_convert_emoji.R
\name{str_convert_emoji}
\alias{str_convert_emoji}
\title{Filter Rows That Contain Markers}
\usage{
str_convert_emoji(
  .str,
  .table_emoji = emoji::emojis,
  .col_emoji = emoji,
  .col_description = name,
  .trans_description = "<{snakecase::to_screaming_snake_case(.desc)}>"
)
}
\arguments{
\item{.str}{...}
}
\value{
\code{str_convert_emoji} - returns a ...
}
\description{
\code{str_convert_emoji} - Convert all Emojis to some ...
}
\examples{
## str_convert_emoji EXAMPLE:

str_convert_emoji(
  "ğŸ˜€ğŸ˜ƒğŸ˜„ğŸ˜ğŸ˜†ğŸ˜…ğŸ˜‚"
)
str_convert_emoji(
  "ğŸ˜€ğŸ˜ƒğŸ˜„ğŸ˜ğŸ˜†ğŸ˜…ğŸ˜‚", .col_description=subgroup
)
str_convert_emoji(
  "ğŸ˜€ğŸ˜ƒğŸ˜„ğŸ˜ğŸ˜†ğŸ˜…ğŸ˜‚", .col_description=group
)
}
