% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_describe_numbers.R
\name{check_num2words}
\alias{check_num2words}
\title{\code{str_describe_numbers} - Replace all cardinal or ordinal numbers in a character vector with generic tags.}
\usage{
check_num2words(.do_install)
}
\arguments{
\item{.do_install}{...}

\item{.str}{...}

\item{.lang}{...}

\item{.ordinal}{...}
}
\value{
\code{str_describe_numbers} - returns a ...

invisible(NULL)
}
\description{
\code{str_describe_numbers} - Replace all cardinal or ordinal numbers in a character vector with generic tags.
}
\examples{
## str_describe_numbers EXAMPLE:

\dontrun{
str_describe_numbers(c(
  "The 2020 United States presidential election",
  "was the 59th quadrennial presidential election",
  "held on Tuesday, November 3, 2020"
))
}
check_num2words

if(FALSE){
  check_num2words(.do_install=TRUE)
}
}
