% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_unify_spacing.R
\name{str_unify_spacing}
\alias{str_unify_spacing}
\title{Filter Rows That Contain Markers}
\usage{
str_unify_spacing(.str)
}
\arguments{
\item{.str}{...}
}
\value{
\code{str_unify_spacing} - returns a ...
}
\description{
\code{str_unify_spacing} - Convert all Emojis to some ...
}
\examples{
## str_unify_spacing EXAMPLE:

str_unify_spacing(
  .str <- "fds  f\n @test_sn fdsfsd #test_ht, <<TAG> test! #dsa"
)
}
